+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-*
+* Use the same margin horizontal in `Item` as the one being used in other components
 
 ## [0.18.0] - 2022-02-21