+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 * Use the same margin horizontal in `Item` as the one being used in other components
+* Align expand buttons in `KeyValues` to the center in tablet screens 
 
 ## [0.18.0] - 2022-02-21