+++ CHANGELOG.md
@@ -9,6 +9,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+*
+
+### Changed
+
+*
+
+### Fixed
+
+*
+
+## [0.19.0] - 2022-02-25
+
+### Added
+
 * Support for `Profile` key value action buttons - [ripe-id-mobile/#58](https://github.com/ripe-tech/ripe-id-mobile/issues/58)
 
 ### Changed